make ShiftCard as a single function to make it easy to handle
